+++ src/types/path.rs
@@ -800,14 +800,14 @@ mod tests {
 
     #[test]
     fn gen_path_2() {
-        let path = Path::from_data(b"M 10 20 l 10 20").unwrap();
+        let path = Path::from_data("M 10 20 l 10 20").unwrap();
         assert_eq!(path.to_string(), "M 10 20 l 10 20");
     }
 
     #[test]
     fn gen_path_3() {
         let path = Path::from_data(
-           b"M 10 20 L 30 40 H 50 V 60 C 70 80 90 100 110 120 \
+            "M 10 20 L 30 40 H 50 V 60 C 70 80 90 100 110 120 \
              S 130 140 150 160 Q 170 180 190 200 T 210 220 \
              A 50 50 30 1 1 230 240 Z").unwrap();
         assert_eq_text!(path.to_string(),
@@ -819,7 +819,7 @@ mod tests {
     #[test]
     fn gen_path_4() {
         let path = Path::from_data(
-           b"m 10 20 l 30 40 h 50 v 60 c 70 80 90 100 110 120 \
+            "m 10 20 l 30 40 h 50 v 60 c 70 80 90 100 110 120 \
              s 130 140 150 160 q 170 180 190 200 t 210 220 \
              a 50 50 30 1 1 230 240 z").unwrap();
         assert_eq_text!(path.to_string(),
@@ -830,7 +830,7 @@ mod tests {
 
     #[test]
     fn gen_path_5() {
-        let path = Path::from_data(b"").unwrap();
+        let path = Path::from_data("").unwrap();
         assert_eq_text!(path.to_string(), "");
     }
 
@@ -849,28 +849,28 @@ mod tests {
     }
 
     test_gen_path_opt!(gen_path_6,
-        b"M 10 20 L 30 40 L 50 60 l 70 80",
-         "M 10 20 L 30 40 50 60 l 70 80",
+        "M 10 20 L 30 40 L 50 60 l 70 80",
+        "M 10 20 L 30 40 50 60 l 70 80",
         remove_duplicated_commands);
 
     test_gen_path_opt!(gen_path_7,
-        b"M 10 20 30 40 50 60",
-         "M 10 20 L 30 40 50 60",
+        "M 10 20 30 40 50 60",
+        "M 10 20 L 30 40 50 60",
         remove_duplicated_commands);
 
     test_gen_path_opt!(gen_path_8,
-        b"M 10 20 L 30 40",
-         "M10 20L30 40",
+        "M 10 20 L 30 40",
+        "M10 20L30 40",
         use_compact_notation);
 
     test_gen_path_opt!(gen_path_9,
-        b"M 10 20 V 30 H 40 V 50 H 60 Z",
-         "M10 20V30H40V50H60Z",
+        "M 10 20 V 30 H 40 V 50 H 60 Z",
+        "M10 20V30H40V50H60Z",
         use_compact_notation);
 
     #[test]
     fn gen_path_10() {
-        let path = Path::from_data(b"M 10 -20 A 5.5 0.3 -4 1 1 0 -0.1").unwrap();
+        let path = Path::from_data("M 10 -20 A 5.5 0.3 -4 1 1 0 -0.1").unwrap();
 
         let mut opt = WriteOptions::default();
         opt.paths.use_compact_notation = true;
@@ -881,44 +881,44 @@ mod tests {
     }
 
     test_gen_path_opt!(gen_path_11,
-        b"M 10-10 a 1 1 0 1 1 -1 1",
-         "M10-10a1 1 0 1 1 -1 1",
+        "M 10-10 a 1 1 0 1 1 -1 1",
+        "M10-10a1 1 0 1 1 -1 1",
         use_compact_notation);
 
     test_gen_path_opt!(gen_path_12,
-        b"M 10-10 a 1 1 0 1 1 0.1 1",
-         "M10-10a1 1 0 1 1 0.1 1",
+        "M 10-10 a 1 1 0 1 1 0.1 1",
+        "M10-10a1 1 0 1 1 0.1 1",
         use_compact_notation);
 
     test_gen_path_opt!(gen_path_13,
-        b"M 10 20 L 30 40 L 50 60 H 10",
-         "M 10 20 30 40 50 60 H 10",
+        "M 10 20 L 30 40 L 50 60 H 10",
+        "M 10 20 30 40 50 60 H 10",
         use_implicit_lineto_commands);
 
     // should be ignored, because of different 'absolute' values
     test_gen_path_opt!(gen_path_14,
-        b"M 10 20 l 30 40 L 50 60",
-         "M 10 20 l 30 40 L 50 60",
+        "M 10 20 l 30 40 L 50 60",
+        "M 10 20 l 30 40 L 50 60",
         use_implicit_lineto_commands);
 
     test_gen_path_opt!(gen_path_15,
-        b"M 10 20 L 30 40 l 50 60 L 50 60",
-         "M 10 20 30 40 l 50 60 L 50 60",
+        "M 10 20 L 30 40 l 50 60 L 50 60",
+        "M 10 20 30 40 l 50 60 L 50 60",
         use_implicit_lineto_commands);
 
     test_gen_path_opt!(gen_path_16,
-        b"M 10 20 L 30 40 l 50 60",
-         "M 10 20 30 40 l 50 60",
+        "M 10 20 L 30 40 l 50 60",
+        "M 10 20 30 40 l 50 60",
         use_implicit_lineto_commands);
 
     test_gen_path_opt!(gen_path_17,
-        b"M 10 20 L 30 40 L 50 60 M 10 20 L 30 40 L 50 60",
-         "M 10 20 30 40 50 60 M 10 20 30 40 50 60",
+        "M 10 20 L 30 40 L 50 60 M 10 20 L 30 40 L 50 60",
+        "M 10 20 30 40 50 60 M 10 20 30 40 50 60",
         use_implicit_lineto_commands);
 
     #[test]
     fn gen_path_18() {
-        let path = Path::from_data(b"M 10 20 L 30 40 L 50 60 M 10 20 L 30 40 L 50 60").unwrap();
+        let path = Path::from_data("M 10 20 L 30 40 L 50 60 M 10 20 L 30 40 L 50 60").unwrap();
 
         let mut opt = WriteOptions::default();
         opt.paths.use_implicit_lineto_commands = true;
@@ -929,7 +929,7 @@ mod tests {
 
     #[test]
     fn gen_path_19() {
-        let path = Path::from_data(b"M 10.00000001 10.00002 L 20.123456789 20.9999999").unwrap();
+        let path = Path::from_data("M 10.00000001 10.00002 L 20.123456789 20.9999999").unwrap();
 
         let mut opt = WriteOptions::default();
 
@@ -967,51 +967,51 @@ mod to_absolute {
     }
 
     test!(line_to,
-          b"m 10 20 l 20 20",
-           "M 10 20 L 30 40");
+          "m 10 20 l 20 20",
+          "M 10 20 L 30 40");
 
     test!(close_path,
-          b"m 10 20 l 20 20 z",
-           "M 10 20 L 30 40 Z");
+          "m 10 20 l 20 20 z",
+          "M 10 20 L 30 40 Z");
 
     // test to check that libsvgparser parses implicit MoveTo as LineTo
     test!(implicit_line_to,
-          b"m 10 20 20 20",
-           "M 10 20 L 30 40");
+          "m 10 20 20 20",
+          "M 10 20 L 30 40");
 
     test!(hline_vline,
-          b"m 10 20 v 10 h 10 l 10 10",
-           "M 10 20 V 30 H 20 L 30 40");
+          "m 10 20 v 10 h 10 l 10 10",
+          "M 10 20 V 30 H 20 L 30 40");
 
     test!(curve,
-          b"m 10 20 c 10 10 10 10 10 10",
-           "M 10 20 C 20 30 20 30 20 30");
+          "m 10 20 c 10 10 10 10 10 10",
+          "M 10 20 C 20 30 20 30 20 30");
 
     test!(move_to_1,
-          b"m 10 20 l 10 10 m 10 10 l 10 10",
-           "M 10 20 L 20 30 M 30 40 L 40 50");
+          "m 10 20 l 10 10 m 10 10 l 10 10",
+          "M 10 20 L 20 30 M 30 40 L 40 50");
 
     test!(move_to_2,
-          b"m 10 20 l 10 10 z m 10 10 l 10 10",
-           "M 10 20 L 20 30 Z M 20 30 L 30 40");
+          "m 10 20 l 10 10 z m 10 10 l 10 10",
+          "M 10 20 L 20 30 Z M 20 30 L 30 40");
 
     test!(move_to_3,
-          b"m 10 20 l 10 10 Z m 10 10 l 10 10",
-           "M 10 20 L 20 30 Z M 20 30 L 30 40");
+          "m 10 20 l 10 10 Z m 10 10 l 10 10",
+          "M 10 20 L 20 30 Z M 20 30 L 30 40");
 
     test!(smooth_curve,
-          b"m 10 20 s 10 10 10 10",
-           "M 10 20 S 20 30 20 30");
+          "m 10 20 s 10 10 10 10",
+          "M 10 20 S 20 30 20 30");
 
     test!(quad,
-          b"m 10 20 q 10 10 10 10",
-           "M 10 20 Q 20 30 20 30");
+          "m 10 20 q 10 10 10 10",
+          "M 10 20 Q 20 30 20 30");
 
     test!(arc_mixed,
-          b"M 30 150 a 40 40 0 0 1 65 50 Z m 30 30 A 20 20 0 0 0 125 230 Z \
-            m 40 24 a 20 20 0 0 1 65 50 z",
-           "M 30 150 A 40 40 0 0 1 95 200 Z M 60 180 A 20 20 0 0 0 125 230 Z \
-            M 100 204 A 20 20 0 0 1 165 254 Z");
+          "M 30 150 a 40 40 0 0 1 65 50 Z m 30 30 A 20 20 0 0 0 125 230 Z \
+           m 40 24 a 20 20 0 0 1 65 50 z",
+          "M 30 150 A 40 40 0 0 1 95 200 Z M 60 180 A 20 20 0 0 0 125 230 Z \
+           M 100 204 A 20 20 0 0 1 165 254 Z");
 }
 
 #[cfg(test)]
@@ -1031,48 +1031,48 @@ mod to_relative {
     }
 
     test!(line_to,
-          b"M 10 20 L 30 40",
-           "m 10 20 l 20 20");
+          "M 10 20 L 30 40",
+          "m 10 20 l 20 20");
 
     test!(close_path,
-          b"M 10 20 L 30 40 Z",
-           "m 10 20 l 20 20 z");
+          "M 10 20 L 30 40 Z",
+          "m 10 20 l 20 20 z");
 
     test!(implicit_line_to,
-          b"M 10 20 30 40",
-           "m 10 20 l 20 20");
+          "M 10 20 30 40",
+          "m 10 20 l 20 20");
 
     test!(hline_vline,
-          b"M 10 20 V 30 H 20 L 30 40",
-           "m 10 20 v 10 h 10 l 10 10");
+          "M 10 20 V 30 H 20 L 30 40",
+          "m 10 20 v 10 h 10 l 10 10");
 
     test!(curve,
-          b"M 10 20 C 20 30 20 30 20 30",
-           "m 10 20 c 10 10 10 10 10 10");
+          "M 10 20 C 20 30 20 30 20 30",
+          "m 10 20 c 10 10 10 10 10 10");
 
     test!(move_to_1,
-          b"M 10 20 L 20 30 M 30 40 L 40 50",
-           "m 10 20 l 10 10 m 10 10 l 10 10");
+          "M 10 20 L 20 30 M 30 40 L 40 50",
+          "m 10 20 l 10 10 m 10 10 l 10 10");
 
     test!(move_to_2,
-          b"M 10 20 L 20 30 Z M 20 30 L 30 40",
-           "m 10 20 l 10 10 z m 10 10 l 10 10");
+          "M 10 20 L 20 30 Z M 20 30 L 30 40",
+          "m 10 20 l 10 10 z m 10 10 l 10 10");
 
     test!(move_to_3,
-          b"M 10 20 L 20 30 z M 20 30 L 30 40",
-           "m 10 20 l 10 10 z m 10 10 l 10 10");
+          "M 10 20 L 20 30 z M 20 30 L 30 40",
+          "m 10 20 l 10 10 z m 10 10 l 10 10");
 
     test!(smooth_curve,
-          b"M 10 20 S 20 30 20 30",
-           "m 10 20 s 10 10 10 10");
+          "M 10 20 S 20 30 20 30",
+          "m 10 20 s 10 10 10 10");
 
     test!(quad,
-          b"M 10 20 Q 20 30 20 30",
-           "m 10 20 q 10 10 10 10");
+          "M 10 20 Q 20 30 20 30",
+          "m 10 20 q 10 10 10 10");
 
     test!(arc_mixed,
-          b"M 30 150 a 40 40 0 0 1 65 50 Z m 30 30 A 20 20 0 0 0 125 230 Z \
-            m 40 24 a 20 20 0 0 1 65 50 z",
-           "m 30 150 a 40 40 0 0 1 65 50 z m 30 30 a 20 20 0 0 0 65 50 z \
-            m 40 24 a 20 20 0 0 1 65 50 z");
+          "M 30 150 a 40 40 0 0 1 65 50 Z m 30 30 A 20 20 0 0 0 125 230 Z \
+           m 40 24 a 20 20 0 0 1 65 50 z",
+          "m 30 150 a 40 40 0 0 1 65 50 z m 30 30 a 20 20 0 0 0 65 50 z \
+           m 40 24 a 20 20 0 0 1 65 50 z");
 }